refactor: Extract sso code into features (no-changelog)
