+++ packages/frontend/editor-ui/src/features/sso/views/SettingsSso.test.ts
@@ -2,9 +2,9 @@ import type { OidcConfigDto, SamlPreferences } from '@n8n/api-types';
 import { createTestingPinia } from '@pinia/testing';
 import { within, waitFor } from '@testing-library/vue';
 import { type MockedStore, mockedStore } from '@/__tests__/utils';
-import SettingsSso from '@/views/SettingsSso.vue';
+import SettingsSso from './SettingsSso.vue';
 import userEvent from '@testing-library/user-event';
-import { useSSOStore, SupportedProtocols } from '@/stores/sso.store';
+import { useSSOStore, SupportedProtocols } from '../sso.store';
 import { createComponentRenderer } from '@/__tests__/render';
 import { usePageRedirectionHelper } from '@/composables/usePageRedirectionHelper';
 import type { SamlPreferencesExtractedData } from '@n8n/rest-api-client/api/sso';
