+++ packages/frontend/editor-ui/src/views/SettingsPersonalView.vue
@@ -21,7 +21,7 @@ import { createFormEventBus } from '@n8n/design-system/utils';
 import type { MfaModalEvents } from '@/event-bus/mfa';
 import { promptMfaCodeBus } from '@/event-bus/mfa';
 import type { BaseTextKey } from '@n8n/i18n';
-import { useSSOStore } from '@/stores/sso.store';
+import { useSSOStore } from '@/features/sso/sso.store';
 import type { ConfirmPasswordModalEvents } from '@/components/ConfirmPasswordModal/confirm-password.event-bus';
 import { confirmPasswordEventBus } from '@/components/ConfirmPasswordModal/confirm-password.event-bus';
